This App is developed to check whether the input file has JUMBF and which type of JUMBF Description box is present in file.
This software helps users to distinguish between JPEG 1 file and JPEG file with JUMBF as well as allows users to check which JUMBF Type is present in the input file. This is beneficial in case users do not have decoders of different types, they can use this software to check whether the file has JUMBF or not and which JUMBF is there.

Co-Authored-By: Faizullah <faizullah877@gmail.com>